Adds UpdateMediaEntry Call

+++ Semesterprojekt/Semesterprojekt/Controllers/MediaEntryController.cs
@@ -6,6 +6,7 @@ using Semesterprojekt.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Authentication;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
@@ -92,8 +93,29 @@ namespace Semesterprojekt.Controllers
         {
             int mediaEntryId = Int32.Parse(mediaEntryIdString);
             string username = HttpUtility.ValidateJwtToken(token);
+            if (username == "")
+            {
+                throw new InvalidCredentialException("Invalid Token!");
+            }
             User user = userService.GetUserByUsername(username);
             mediaEntryService.DeleteMediaEntry(mediaEntryId, user.UserId);
+        }
+
+        public void UpdateMedia(string token, string mediaEntryIdString, string requestBody)
+        {
+            int mediaEntryId = Int32.Parse(mediaEntryIdString);
+            string username = HttpUtility.ValidateJwtToken(token);
+            if (username == "")
+            {
+                throw new InvalidCredentialException("Invalid Token!");
+            }
+            User user = userService.GetUserByUsername(username);
+            MediaEntryDTO mediaEntryDTO = JsonSerializer.Deserialize<MediaEntryDTO>(requestBody);
+            if (mediaEntryDTO == null)
+            {
+                throw new InvalidRequestBodyException("Invalid Request!");
+            }
+            mediaEntryService.UpdateMediaEntry(user.UserId, mediaEntryId, mediaEntryDTO);
         }
     }
 }
